--- .env copy
@@ -1,1 +1,0 @@
-MONGO_DEV_URI=mongodb://localhost:27017/emoji-hero
