+++ .gitignore
@@ -17,6 +17,7 @@
 .env.development.local
 .env.test.local
 .env.production.local
+.env copy
 
 npm-debug.log*
 yarn-debug.log*
