+++ server/server.js
@@ -1,32 +1,33 @@
-const express = require('express');
-const { ApolloServer } = require('apollo-server-express');
+// this is our server file
+// this will be our express server
+import express from 'express'
+import connection from './connection/index.js'
 
-const { typeDefs, resolvers } = require('./schemas');
-const db = require('./config/connection');
 
-const PORT = process.env.PORT || 3001;
+// apollo graphql server (graphql/apollo)
+// set up typeDefs (graphql/apollo)
+// set up resolvers for all matching typeDefs (graphql/apollo)
+// set up models for our database (mongoose)
+// set up the actual database connection (mongoose)
+// make middleware for certain use cases/ie debugging (express middleware (app.use(middleware)))
+
+import { ApolloServer } from 'apollo-server-express'
+import { resolvers } from './resolvers/index.js'
+import { typeDefs } from './typeDefs/index.js'
+const PORT = 3001
+const app = express()
+app.use(express.json())
+app.use(express.urlencoded({ extended: true }))
+
 const server = new ApolloServer({
-  typeDefs,
-  resolvers
-});
+    typeDefs,
+    resolvers
+})
+await server.start()
 
-const app = express();
-
-app.use(express.urlencoded({ extended: false }));
-app.use(express.json());
- 
-// Create a new instance of an Apollo server with the GraphQL schema
-const startApolloServer = async (typeDefs, resolvers) => {
-  await server.start();
-  server.applyMiddleware({ app });
-  
-  db.once('open', () => {
+server.applyMiddleware({ app })
+if (connection) {
     app.listen(PORT, () => {
-      console.log(`API server running on port ${PORT}!`);
-      console.log(`Use GraphQL at http://localhost:${PORT}${server.graphqlPath}`);
+        console.log("listening on port " + PORT)
     })
-  })
-  };
-  
-  // Call the async function to start the server
-  startApolloServer(typeDefs, resolvers);
+}
